+++ src/main/java/io/cucumber/pro/JsonReporter.java
@@ -7,13 +7,10 @@ import cucumber.api.event.TestRunFinished;
 import cucumber.api.formatter.Formatter;
 import cucumber.runtime.formatter.PluginFactory;
 import io.cucumber.pro.config.Config;
-import io.cucumber.pro.documentation.DocumentationPublisher;
-import io.cucumber.pro.documentation.DocumentationPublisherFactory;
+import io.cucumber.pro.environment.CIEnvironment;
 import io.cucumber.pro.environment.EnvFilter;
 import io.cucumber.pro.results.ResultsPublisher;
 import io.cucumber.pro.results.ResultsPublisherFactory;
-import io.cucumber.pro.revision.RevisionProvider;
-import io.cucumber.pro.revision.RevisionProviderFactory;
 
 import java.io.File;
 import java.io.IOException;
@@ -23,32 +20,32 @@ public class JsonReporter implements Formatter {
 
     private static final Config CONFIG = ConfigFactory.create();
     private static final Logger LOGGER = new Logger.SystemLogger(CONFIG);
-    private static final RevisionProvider REVISION_PROVIDER = RevisionProviderFactory.create(CONFIG, LOGGER);
+    private static final CIEnvironment CI_ENVIRONMENT = CIEnvironment.detect(CONFIG);
     private final Formatter jsonFormatter;
     private final File jsonFile;
     private final ResultsPublisher resultsPublisher;
     private final String profileName;
     private final Logger logger;
     private final Map<String, String> env;
-    private final DocumentationPublisher documentationPublisher;
     private final Config config;
 
     JsonReporter(
-            DocumentationPublisher documentationPublisher,
             ResultsPublisher resultsPublisher,
             String profileName,
             Config config,
             Logger logger,
-            RevisionProvider revisionProvider,
+            CIEnvironment ciEnvironment,
             Map<String, String> env
     ) {
-        this.documentationPublisher = documentationPublisher;
         this.resultsPublisher = resultsPublisher;
         this.profileName = profileName;
         this.config = config;
         this.logger = logger;
         this.env = new EnvFilter(config).filter(env);
-        this.env.put("cucumber_pro_git_branch", revisionProvider.getBranch());
+
+        if (ciEnvironment != null) {
+            this.env.put("cucumber_pro_git_branch", ciEnvironment.getBranch(config));
+        }
 
         try {
             jsonFile = File.createTempFile("cucumber-json", ".json");
@@ -61,16 +58,15 @@ public class JsonReporter implements Formatter {
 
     public JsonReporter(String profileName) {
         this(
-                DocumentationPublisherFactory.create(CONFIG, LOGGER),
                 ResultsPublisherFactory.create(
                         CONFIG,
                         LOGGER,
-                        REVISION_PROVIDER
+                        CI_ENVIRONMENT
                 ),
                 profileName,
                 CONFIG,
                 LOGGER,
-                REVISION_PROVIDER,
+                CI_ENVIRONMENT,
                 System.getenv()
         );
     }
@@ -101,7 +97,6 @@ public class JsonReporter implements Formatter {
                     @Override
                     public void receive(TestRunFinished event) {
                         JsonReporter.this.logger.log(Logger.Level.DEBUG, "Cucumber Pro config:\n\n%s", JsonReporter.this.config.toYaml("cucumberpro"));
-                        JsonReporter.this.documentationPublisher.publish();
                         JsonReporter.this.resultsPublisher.publish(jsonFile, env, profileName);
                     }
                 });