+++ src/main/java/io/cucumber/pro/JsonReporter12.java
@@ -2,13 +2,10 @@ package io.cucumber.pro;
 
 import gherkin.formatter.JSONFormatter;
 import io.cucumber.pro.config.Config;
-import io.cucumber.pro.documentation.DocumentationPublisher;
-import io.cucumber.pro.documentation.DocumentationPublisherFactory;
+import io.cucumber.pro.environment.CIEnvironment;
 import io.cucumber.pro.environment.EnvFilter;
 import io.cucumber.pro.results.ResultsPublisher;
 import io.cucumber.pro.results.ResultsPublisherFactory;
-import io.cucumber.pro.revision.RevisionProvider;
-import io.cucumber.pro.revision.RevisionProviderFactory;
 
 import java.io.File;
 import java.io.FileWriter;
@@ -19,7 +16,7 @@ public class JsonReporter12 extends JSONFormatter {
 
     private static final Config CONFIG = ConfigFactory.create();
     private static final Logger LOGGER = new Logger.SystemLogger(CONFIG);
-    private static final RevisionProvider REVISION_PROVIDER = RevisionProviderFactory.create(CONFIG, LOGGER);
+    private static final CIEnvironment CI_ENVIRONMENT = CIEnvironment.detect(CONFIG);
 
     private static final File jsonFile;
 
@@ -36,41 +33,39 @@ public class JsonReporter12 extends JSONFormatter {
     private final Logger logger;
     private final ResultsPublisher resultsPublisher;
     private final String profileName;
-    private final DocumentationPublisher documentationPublisher;
     private final Map<String, String> env;
 
     JsonReporter12(
-            DocumentationPublisher documentationPublisher,
             ResultsPublisher resultsPublisher,
             String profileName,
             Config config,
             Logger logger,
-            RevisionProvider revisionProvider,
+            CIEnvironment ciEnvironment,
             Map<String, String> env
     ) throws IOException {
         super(new FileWriter(jsonFile));
-        this.documentationPublisher = documentationPublisher;
         this.resultsPublisher = resultsPublisher;
         this.profileName = profileName;
         this.config = config;
         this.logger = logger;
 
         this.env = new EnvFilter(config).filter(env);
-        this.env.put("cucumber_pro_git_branch", revisionProvider.getBranch());
+        if (ciEnvironment != null) {
+            this.env.put("cucumber_pro_git_branch", ciEnvironment.getBranch(config));
+        }
     }
 
     JsonReporter12(String profileName) throws IOException {
         this(
-                DocumentationPublisherFactory.create(CONFIG, LOGGER),
                 ResultsPublisherFactory.create(
                         CONFIG,
                         LOGGER,
-                        REVISION_PROVIDER
+                        CI_ENVIRONMENT
                 ),
                 profileName,
                 CONFIG,
                 LOGGER,
-                REVISION_PROVIDER,
+                CI_ENVIRONMENT,
                 System.getenv()
         );
     }
@@ -84,7 +79,6 @@ public class JsonReporter12 extends JSONFormatter {
     public void close() {
         super.close();
         logger.log(Logger.Level.DEBUG, "Cucumber Pro config:\n\n%s", config.toYaml("cucumberpro"));
-        this.documentationPublisher.publish();
         this.resultsPublisher.publish(jsonFile, env, profileName);
     }
 }