+++ src/main/java/io/cucumber/pro/documentation/DocumentationPublisher.java
@@ -1,5 +1,0 @@
-package io.cucumber.pro.documentation;
-
-public interface DocumentationPublisher {
-    void publish();
-}