+++ src/main/java/io/cucumber/pro/documentation/NullDocumentationPublisher.java
@@ -1,7 +1,0 @@
-package io.cucumber.pro.documentation;
-
-public class NullDocumentationPublisher implements DocumentationPublisher {
-    @Override
-    public void publish() {
-    }
-}