+++ src/main/java/io/cucumber/pro/results/HTTPResultsPublisher.java
@@ -51,7 +51,7 @@ class HTTPResultsPublisher implements ResultsPublisher {
     HTTPResultsPublisher(String url, Config config, Logger logger) {
         this.url = url;
         this.config = config;
-        this.authToken = config.getString(Keys.CUCUMBERPRO_RESULTS_TOKEN);
+        this.authToken = config.getString(Keys.CUCUMBERPRO_TOKEN);
         this.logger = logger;
     }
 
@@ -84,9 +84,9 @@ class HTTPResultsPublisher implements ResultsPublisher {
 
                 String suggestion = "";
                 if (statusCode == 401)
-                    suggestion = String.format("You need to define %s", Keys.CUCUMBERPRO_RESULTS_TOKEN);
+                    suggestion = String.format("You need to define %s", Keys.CUCUMBERPRO_TOKEN);
                 if (statusCode == 403)
-                    suggestion = String.format("You need to change the value of %s", Keys.CUCUMBERPRO_RESULTS_TOKEN);
+                    suggestion = String.format("You need to change the value of %s", Keys.CUCUMBERPRO_TOKEN);
 
                 String message = String.format(
                         "Failed to publish results to Cucumber Pro URL: %s, Status: %s\n%s\n%s",