+++ src/main/java/io/cucumber/pro/revision/RevisionProvider.java
@@ -1,7 +1,0 @@
-package io.cucumber.pro.revision;
-
-public interface RevisionProvider {
-    String getRevision();
-
-    String getBranch();
-}